requirements.txt
aiogram
pytz
httpx>=0.24.0
google-genai
groq
openai
hltv
sportsipy
python-dotenv
beautifulsoup4
lxml
ddgs
tavily-python
exa-py
# PostgreSQL (optional, for production)
# psycopg2-binary